+++ R/drop_na_if.R
@@ -72,7 +72,7 @@ if (all(!is.null(keep_columns), !is.null(target_columns))){
   
 if (!is.null(keep_columns)){
   grouped_dfr %>% 
-    dplyr::filter(dplyr::across(-c(keep_columns),
+    dplyr::filter(dplyr::across(-all_of(c(keep_columns)),
                   ~!switches(mean(is.na(.)) * 100,
                                sign=sign,
                 percent_na=percent_na))) %>% 
@@ -83,7 +83,7 @@ if (!is.null(keep_columns)){
     # if we provide a specific column, only drop at these
     
     grouped_dfr %>%   
-    dplyr::filter(dplyr::across(target_columns,
+    dplyr::filter(dplyr::across(all_of(target_columns),
                       ~!switches(mean(is.na(.)) * 100,
                         sign=sign,
                 percent_na=percent_na))) %>%